backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.4",
    "langchain>=1.1.3",
    "langchain-anthropic>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.0",
    "langgraph>=1.0.5",
    "pinecone>=8.1.1",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "sentry-sdk>=2.57.0",
    "uvicorn[standard]>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.12.0",
    "responses>=0.25.0",
]
